# .migration-backup/lib/api-zod/src/generated/types/adminCustomRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Tools4Biz marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminCustomRequest {
  id: number;
  /** @nullable */
  name?: string | null;
  email: string;
  /** @nullable */
  businessType?: string | null;
  description: string;
  /** @nullable */
  budget?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  status: string;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  paymentAmount?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}