# .migration-backup/lib/api-zod/src/generated/types/adminSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Tools4Biz marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminSummary {
  totalTools: number;
  totalWaitlist: number;
  totalCustomRequests: number;
  newCustomRequests: number;
  paidRequests: number;
  totalRevenue: number;
}