# .migration-backup/lib/api-zod/src/generated/types/adminWaitlistEntry.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Tools4Biz marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminWaitlistEntry {
  id: number;
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  toolId?: number | null;
  /** @nullable */
  toolName?: string | null;
  /** @nullable */
  message?: string | null;
  createdAt: string;
}