# .migration-backup/lib/api-zod/src/generated/types/toolUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Tools4Biz marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { ToolUpdateStatus } from './toolUpdateStatus';

export interface ToolUpdate {
  name?: string;
  description?: string;
  category?: string;
  price?: number;
  currency?: string;
  status?: ToolUpdateStatus;
  featured?: boolean;
  emoji?: string;
  tagline?: string;
  features?: string[];
}