# .migration-backup/lib/api-zod/src/generated/types/waitlistEntry.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Tools4Biz marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface WaitlistEntry {
  id: number;
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  toolId?: number | null;
  createdAt: string;
}