# .migration-backup/lib/api-zod/src/generated/types/waitlistInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Tools4Biz marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface WaitlistInput {
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  toolId?: number | null;
  /** @nullable */
  message?: string | null;
}